refactor(notifications): Drop the per-kind filter chips
The notification feed carried a row of nine toggle chips (All plus one per
kind) backed by a persisted selection. It cost real complexity and bought
very little: the feed is short, free-text search already narrows it, and the
notifications page's stat cards already break volume down by kind. Nobody
needs to hide cron notifications to find an approval in a list of twelve.

It was also actively harmful. The selection persisted as an EXPLICIT list of
kinds while the feed treated "every known kind is selected" as a special
include-unknown-kinds state. So adding a kind turned a stored full set into a
partial one, which flipped filtering to strict and hid the new kind for every
existing install -- with no chip that brought it back until the user happened
to toggle something. That bug shipped once already (skill notifications, fixed
in #1473 by versioning the storage key and migrating). Removing the filter
makes the whole failure mode impossible by construction instead of managing it
with a migration.

What goes:
  - CATEGORIES, Category, KIND_KEYS, LEGACY_KIND_KEYS, parseKinds,
    loadActiveKinds, KINDS_STORAGE_KEY, LEGACY_KINDS_STORAGE_KEY
  - the activeKinds state, its persist effect, toggleCategory, allActive /
    noneActive, and the chips row
  - the now-unreachable "no categories selected" empty state

What stays:
  - KIND_META / DEFAULT_META -- the per-row icon, badge and border colour are
    not a filter
  - the free-text search box
  - the muted-channel disclosure. It is NOT a kind filter: it reveals rows
    that are otherwise unreachable under mute-keeps-history semantics, so
    deleting it would either restore the clutter mute exists to prevent or
    make muted notifications permanently invisible. Its row now renders only
    when something is actually silenced, so a normal feed shows no chrome at
    all.

Its title attribute went away with the chips row rather than being ported: it
was a raw English template literal that also concatenated a literal plural
("notification" + "s"), the exact shape the [plurals] gate forbids. The button
keeps aria-pressed and an already-translated visible label, so nothing is lost
and no new key was needed.

Ten catalog keys are now referenced nowhere and are removed from every locale
(en-XA regenerated, one orphaned en.context.json entry dropped). Three of them
were chip-only lookalikes worth naming, because a substring grep says they are
still in use: kind_cron, kind_hooks and kind_tasks are distinct from the badge
keys kind_cron_job, kind_webhook and kind_task_runner. The dead-key scan is
authoritative here, not grep.

NotificationsPage.filter.test.tsx and notifMeta.activeKinds.test.ts are deleted
-- all eighteen tests drove the removed machinery. Two of their assertions
covered a real invariant rather than the chips, and those are preserved in a new
NotificationsPage.noFilter.test.tsx: a notification of any kind renders,
including one whose kind the frontend does not know. That case used to be
conditional on the all-selected state, which is precisely how it broke.
