+++ website/src/i18n/locales/bn.json
@@ -4276,17 +4276,13 @@
         "unknown_date": "অজানা তারিখ",
         "yesterday": "গতকাল",
         "kind_agent": "এজেন্ট",
-        "kind_all": "সব",
         "kind_approval": "অনুমোদন",
-        "kind_cron": "Cron",
         "kind_cron_job": "Cron জব",
         "kind_heartbeat": "Heartbeat",
-        "kind_hooks": "Hooks",
         "kind_notification": "নোটিফিকেশন",
         "kind_skills": "স্কিল",
         "kind_subagent": "সাবএজেন্ট",
         "kind_task_runner": "টাস্ক রানার",
-        "kind_tasks": "টাস্ক",
         "kind_webhook": "ওয়েবহুক"
       },
       "notificationDetailPanel": {
@@ -4313,24 +4309,18 @@
         "all": "সব",
         "approve": "অনুমোদন করুন",
         "clear": "সাফ করুন",
-        "clear_all_filters": "সব ফিল্টার সাফ করো",
         "clear_all_notifications": "সব নোটিফিকেশন সাফ করবেন?",
         "dismiss_notification": "নোটিফিকেশন বাতিল করুন",
-        "filter_notifications_by_kind": "ধরন অনুযায়ী নোটিফিকেশন ফিল্টার করুন",
         "first_notification_from": "প্রথম নোটিফিকেশন, উৎস:",
-        "hide_muted_channel_notifications": "মিউট করা চ্যানেলের নোটিফিকেশন লুকাও",
         "keep": "চালু রাখুন",
         "keep_receiving_these": ". এগুলো পেতে থাকতে চান?",
         "mark_all_as_read": "সবগুলো পঠিত হিসেবে চিহ্নিত করুন",
         "mute_channel": "চ্যানেল মিউট করুন",
         "muted_2": "মিউট করা",
         "muted_count": "মিউট করা ({{count}})",
-        "no_categories_selected_click_a_category_above": "কোনো ক্যাটাগরি নির্বাচন করা হয়নি — উপরে একটি ক্যাটাগরিতে ক্লিক করো",
         "no_notifications": "কোনো নোটিফিকেশন নেই",
         "reject": "প্রত্যাখ্যান করুন",
         "search": "খুঁজুন…",
-        "toggle_kind": "{{kind}} টগল করুন",
-        "select_all_categories": "সব ক্যাটাগরি নির্বাচন করো",
         "show_less": "কম দেখাও",
         "try_a_different_search": "অন্য কিছু দিয়ে খুঁজে দেখো"
       }
