+++ website/src/i18n/locales/de.json
@@ -4276,17 +4276,13 @@
         "unknown_date": "Unbekanntes Datum",
         "yesterday": "Gestern",
         "kind_agent": "Agent",
-        "kind_all": "Alle",
         "kind_approval": "Genehmigung",
-        "kind_cron": "Cron",
         "kind_cron_job": "Cron-Job",
         "kind_heartbeat": "Heartbeat",
-        "kind_hooks": "Hooks",
         "kind_notification": "Benachrichtigung",
         "kind_skills": "Fähigkeiten",
         "kind_subagent": "Subagent",
         "kind_task_runner": "Task Runner",
-        "kind_tasks": "Aufgaben",
         "kind_webhook": "Webhook"
       },
       "notificationDetailPanel": {
@@ -4313,24 +4309,18 @@
         "all": "Alle",
         "approve": "Genehmigen",
         "clear": "Leeren",
-        "clear_all_filters": "Alle Filter löschen",
         "clear_all_notifications": "Alle Benachrichtigungen löschen?",
         "dismiss_notification": "Benachrichtigung schließen",
-        "filter_notifications_by_kind": "Benachrichtigungen nach Art filtern",
         "first_notification_from": "Erste Benachrichtigung von",
-        "hide_muted_channel_notifications": "Benachrichtigungen stummgeschalteter Kanäle ausblenden",
         "keep": "Behalten",
         "keep_receiving_these": ". Diese weiterhin erhalten?",
         "mark_all_as_read": "Alle als gelesen markieren",
         "mute_channel": "Kanal stummschalten",
         "muted_2": "stummgeschaltet",
         "muted_count": "Stummgeschaltet ({{count}})",
-        "no_categories_selected_click_a_category_above": "Keine Kategorien ausgewählt — klicke oben auf eine Kategorie",
         "no_notifications": "Keine Benachrichtigungen",
         "reject": "Ablehnen",
         "search": "Suchen…",
-        "toggle_kind": "{{kind}} umschalten",
-        "select_all_categories": "Alle Kategorien auswählen",
         "show_less": "Weniger anzeigen",
         "try_a_different_search": "Versuche eine andere Suche"
       }
