+++ website/src/i18n/locales/en-XA.json
@@ -4110,7 +4110,6 @@
         "clear": "[Çĺèàŕ ········]",
         "clear_all_notifications": "[Çĺèàŕ àĺĺ ñøţìƒìçàţìøñş? ·················]",
         "dismiss_notification": "[Ðìşɱìşş ñøţìƒìçàţìøñ ··················]",
-        "filter_notifications_by_kind": "[Ƒìĺţèŕ ñøţìƒìçàţìøñş ƀý ķìñð ····················]",
         "first_notification_from": "[Ƒìŕşţ ñøţìƒìçàţìøñ ƒŕøɱ ················]",
         "keep": "[Ķèèþ ······]",
         "keep_receiving_these": "[. Ķèèþ ŕèçèìṽìñğ ţĥèşè? ················]",
@@ -4121,14 +4120,9 @@
         "reject": "[Ŕèĵèçţ ·········]",
         "search": "[Şèàŕçĥ… ···········]",
         "activity_will_appear_here": "[Àçţìṽìţý ẁìĺĺ àþþèàŕ ĥèŕè ··················]",
-        "clear_all_filters": "[Çĺèàŕ àĺĺ ƒìĺţèŕş ···············]",
-        "hide_muted_channel_notifications": "[Ĥìðè ɱùţèð-çĥàññèĺ ñøţìƒìçàţìøñş ················]",
         "muted_count": "[Ṁùţèð ({{count}}) ············]",
-        "no_categories_selected_click_a_category_above": "[Ñø çàţèğøŕìèş şèĺèçţèð — çĺìçķ à çàţèğøŕý àƀøṽè ························]",
-        "select_all_categories": "[Şèĺèçţ àĺĺ çàţèğøŕìèş ···············]",
         "show_less": "[Şĥøẁ ĺèşş ··············]",
-        "try_a_different_search": "[Ţŕý à ðìƒƒèŕèñţ şèàŕçĥ ···············]",
-        "toggle_kind": "[Ţøğğĺè {{kind}} ···········]"
+        "try_a_different_search": "[Ţŕý à ðìƒƒèŕèñţ şèàŕçĥ ···············]"
       },
       "notifMeta": {
         "this_week": "[Ţĥìş Ẁèèķ ··············]",
@@ -4136,17 +4130,13 @@
         "unknown_date": "[Ùñķñøẁñ ðàţè ···········]",
         "yesterday": "[Ýèşţèŕðàý ··············]",
         "kind_agent": "[Àğèñţ ········]",
-        "kind_all": "[Àĺĺ ·····]",
         "kind_approval": "[Àþþŕøṽàĺ ············]",
-        "kind_cron": "[Çŕøñ ······]",
         "kind_cron_job": "[Çŕøñ Ĵøƀ ············]",
         "kind_heartbeat": "[Ĥèàŕţƀèàţ ··············]",
-        "kind_hooks": "[Ĥøøķş ········]",
         "kind_notification": "[Ñøţìƒìçàţìøñ ···········]",
         "kind_skills": "[Şķìĺĺş ·········]",
         "kind_subagent": "[Şùƀàğèñţ ············]",
         "kind_task_runner": "[Ţàşķ Ŕùññèŕ ··········]",
-        "kind_tasks": "[Ţàşķş ········]",
         "kind_webhook": "[Ẁèƀĥøøķ ···········]"
       }
     },
