+++ website/src/i18n/locales/en.json
@@ -3447,7 +3447,6 @@
         "clear": "Clear",
         "clear_all_notifications": "Clear all notifications?",
         "dismiss_notification": "Dismiss notification",
-        "filter_notifications_by_kind": "Filter notifications by kind",
         "first_notification_from": "First notification from",
         "keep": "Keep",
         "keep_receiving_these": ". Keep receiving these?",
