+++ website/src/i18n/locales/en.manual.json
@@ -997,29 +997,20 @@
         "unknown_date": "Unknown date",
         "yesterday": "Yesterday",
         "kind_agent": "Agent",
-        "kind_all": "All",
         "kind_approval": "Approval",
-        "kind_cron": "Cron",
         "kind_cron_job": "Cron Job",
         "kind_heartbeat": "Heartbeat",
-        "kind_hooks": "Hooks",
         "kind_notification": "Notification",
         "kind_skills": "Skills",
         "kind_subagent": "Subagent",
         "kind_task_runner": "Task Runner",
-        "kind_tasks": "Tasks",
         "kind_webhook": "Webhook"
       },
       "notificationFeed": {
         "activity_will_appear_here": "Activity will appear here",
-        "clear_all_filters": "Clear all filters",
-        "hide_muted_channel_notifications": "Hide muted-channel notifications",
         "muted_count": "Muted ({{count}})",
-        "no_categories_selected_click_a_category_above": "No categories selected — click a category above",
-        "select_all_categories": "Select all categories",
         "show_less": "Show less",
-        "try_a_different_search": "Try a different search",
-        "toggle_kind": "Toggle {{kind}}"
+        "try_a_different_search": "Try a different search"
       }
     },
     "onboardingFlow": {
