+++ website/src/i18n/locales/es.json
@@ -4331,17 +4331,13 @@
         "unknown_date": "Fecha desconocida",
         "yesterday": "Ayer",
         "kind_agent": "Agente",
-        "kind_all": "Todos",
         "kind_approval": "Aprobación",
-        "kind_cron": "Cron",
         "kind_cron_job": "Trabajo de cron",
         "kind_heartbeat": "Heartbeat",
-        "kind_hooks": "Hooks",
         "kind_notification": "Notificación",
         "kind_skills": "Habilidades",
         "kind_subagent": "Subagente",
         "kind_task_runner": "Ejecutor de tareas",
-        "kind_tasks": "Tareas",
         "kind_webhook": "Webhook"
       },
       "notificationDetailPanel": {
@@ -4368,24 +4364,18 @@
         "all": "Todas",
         "approve": "Aprobar",
         "clear": "Borrar",
-        "clear_all_filters": "Borrar todos los filtros",
         "clear_all_notifications": "¿Borrar todas las notificaciones?",
         "dismiss_notification": "Descartar notificación",
-        "filter_notifications_by_kind": "Filtrar notificaciones por tipo",
         "first_notification_from": "Primera notificación de",
-        "hide_muted_channel_notifications": "Ocultar las notificaciones de canales silenciados",
         "keep": "Mantener",
         "keep_receiving_these": ". ¿Seguir recibiéndolas?",
         "mark_all_as_read": "Marcar todas como leídas",
         "mute_channel": "Silenciar canal",
         "muted_2": "silenciada",
         "muted_count": "Silenciadas ({{count}})",
-        "no_categories_selected_click_a_category_above": "No hay categorías seleccionadas — haz clic en una categoría de arriba",
         "no_notifications": "Sin notificaciones",
         "reject": "Rechazar",
         "search": "Buscar…",
-        "toggle_kind": "Alternar {{kind}}",
-        "select_all_categories": "Seleccionar todas las categorías",
         "show_less": "Mostrar menos",
         "try_a_different_search": "Prueba con otra búsqueda"
       }
