+++ website/src/i18n/locales/fr.json
@@ -4331,17 +4331,13 @@
         "unknown_date": "Date inconnue",
         "yesterday": "Hier",
         "kind_agent": "Agent",
-        "kind_all": "Tous",
         "kind_approval": "Approbation",
-        "kind_cron": "Cron",
         "kind_cron_job": "Tâche cron",
         "kind_heartbeat": "Heartbeat",
-        "kind_hooks": "Hooks",
         "kind_notification": "Notification",
         "kind_skills": "Compétences",
         "kind_subagent": "Sous-agent",
         "kind_task_runner": "Exécuteur de tâches",
-        "kind_tasks": "Tâches",
         "kind_webhook": "Webhook"
       },
       "notificationDetailPanel": {
@@ -4368,24 +4364,18 @@
         "all": "Tout",
         "approve": "Approuver",
         "clear": "Effacer",
-        "clear_all_filters": "Effacer tous les filtres",
         "clear_all_notifications": "Effacer toutes les notifications ?",
         "dismiss_notification": "Ignorer la notification",
-        "filter_notifications_by_kind": "Filtrer les notifications par type",
         "first_notification_from": "Première notification de",
-        "hide_muted_channel_notifications": "Masquer les notifications des canaux en sourdine",
         "keep": "Conserver",
         "keep_receiving_these": ". Continuer à les recevoir ?",
         "mark_all_as_read": "Tout marquer comme lu",
         "mute_channel": "Mettre le canal en sourdine",
         "muted_2": "en sourdine",
         "muted_count": "En sourdine ({{count}})",
-        "no_categories_selected_click_a_category_above": "Aucune catégorie sélectionnée — clique sur une catégorie ci-dessus",
         "no_notifications": "Aucune notification",
         "reject": "Refuser",
         "search": "Rechercher…",
-        "toggle_kind": "Basculer {{kind}}",
-        "select_all_categories": "Sélectionner toutes les catégories",
         "show_less": "Afficher moins",
         "try_a_different_search": "Essaie une autre recherche"
       }
