+++ website/src/i18n/locales/hi.json
@@ -4276,17 +4276,13 @@
         "unknown_date": "अज्ञात तारीख़",
         "yesterday": "कल",
         "kind_agent": "एजेंट",
-        "kind_all": "सभी",
         "kind_approval": "मंज़ूरी",
-        "kind_cron": "Cron",
         "kind_cron_job": "Cron जॉब",
         "kind_heartbeat": "हार्टबीट",
-        "kind_hooks": "हुक",
         "kind_notification": "सूचना",
         "kind_skills": "स्किल्स",
         "kind_subagent": "सब-एजेंट",
         "kind_task_runner": "टास्क रनर",
-        "kind_tasks": "कार्य",
         "kind_webhook": "वेबहुक"
       },
       "notificationDetailPanel": {
@@ -4313,24 +4309,18 @@
         "all": "सभी",
         "approve": "स्वीकृत करें",
         "clear": "साफ़ करें",
-        "clear_all_filters": "सभी फ़िल्टर हटाएँ",
         "clear_all_notifications": "सभी सूचनाएँ साफ़ करें?",
         "dismiss_notification": "सूचना हटाएँ",
-        "filter_notifications_by_kind": "प्रकार के अनुसार सूचनाएँ फ़िल्टर करें",
         "first_notification_from": "पहली सूचना, प्रेषक",
-        "hide_muted_channel_notifications": "म्यूट किए गए चैनल की सूचनाएँ छिपाएँ",
         "keep": "जारी रखें",
         "keep_receiving_these": "। इन्हें प्राप्त करते रहें?",
         "mark_all_as_read": "सभी को पढ़ा हुआ चिह्नित करें",
         "mute_channel": "चैनल म्यूट करें",
         "muted_2": "म्यूट",
         "muted_count": "म्यूट ({{count}})",
-        "no_categories_selected_click_a_category_above": "कोई श्रेणी नहीं चुनी — ऊपर किसी श्रेणी पर क्लिक करें",
         "no_notifications": "कोई सूचना नहीं",
         "reject": "अस्वीकार करें",
         "search": "खोजें…",
-        "toggle_kind": "{{kind}} टॉगल करें",
-        "select_all_categories": "सभी श्रेणियाँ चुनें",
         "show_less": "कम दिखाएँ",
         "try_a_different_search": "कोई दूसरी खोज आज़माएँ"
       }
