+++ website/src/i18n/locales/it.json
@@ -4331,17 +4331,13 @@
         "unknown_date": "Data sconosciuta",
         "yesterday": "Ieri",
         "kind_agent": "Agente",
-        "kind_all": "Tutti",
         "kind_approval": "Approvazione",
-        "kind_cron": "Cron",
         "kind_cron_job": "Job cron",
         "kind_heartbeat": "Heartbeat",
-        "kind_hooks": "Hook",
         "kind_notification": "Notifica",
         "kind_skills": "Competenze",
         "kind_subagent": "Subagente",
         "kind_task_runner": "Task Runner",
-        "kind_tasks": "Attività",
         "kind_webhook": "Webhook"
       },
       "notificationDetailPanel": {
@@ -4368,24 +4364,18 @@
         "all": "Tutti",
         "approve": "Approva",
         "clear": "Cancella",
-        "clear_all_filters": "Cancella tutti i filtri",
         "clear_all_notifications": "Cancellare tutte le notifiche?",
         "dismiss_notification": "Ignora la notifica",
-        "filter_notifications_by_kind": "Filtra le notifiche per tipo",
         "first_notification_from": "Prima notifica da",
-        "hide_muted_channel_notifications": "Nascondi le notifiche dei canali silenziati",
         "keep": "Mantieni",
         "keep_receiving_these": ". Continuare a riceverli?",
         "mark_all_as_read": "Segna tutto come letto",
         "mute_channel": "Silenzia il canale",
         "muted_2": "silenziato",
         "muted_count": "Silenziate ({{count}})",
-        "no_categories_selected_click_a_category_above": "Nessuna categoria selezionata — clicca una categoria qui sopra",
         "no_notifications": "Nessuna notifica",
         "reject": "Rifiuta",
         "search": "Cerca…",
-        "toggle_kind": "Attiva/disattiva {{kind}}",
-        "select_all_categories": "Seleziona tutte le categorie",
         "show_less": "Mostra meno",
         "try_a_different_search": "Prova una ricerca diversa"
       }
