+++ website/src/i18n/locales/pt.json
@@ -4331,17 +4331,13 @@
         "unknown_date": "Data desconhecida",
         "yesterday": "Ontem",
         "kind_agent": "Agente",
-        "kind_all": "Todos",
         "kind_approval": "Aprovação",
-        "kind_cron": "Cron",
         "kind_cron_job": "Job de cron",
         "kind_heartbeat": "Heartbeat",
-        "kind_hooks": "Hooks",
         "kind_notification": "Notificação",
         "kind_skills": "Habilidades",
         "kind_subagent": "Subagente",
         "kind_task_runner": "Executor de tarefas",
-        "kind_tasks": "Tarefas",
         "kind_webhook": "Webhook"
       },
       "notificationDetailPanel": {
@@ -4368,24 +4364,18 @@
         "all": "Todas",
         "approve": "Aprovar",
         "clear": "Limpar",
-        "clear_all_filters": "Limpar todos os filtros",
         "clear_all_notifications": "Limpar todas as notificações?",
         "dismiss_notification": "Descartar notificação",
-        "filter_notifications_by_kind": "Filtrar notificações por tipo",
         "first_notification_from": "Primeira notificação de",
-        "hide_muted_channel_notifications": "Ocultar notificações de canais silenciados",
         "keep": "Manter",
         "keep_receiving_these": ". Continuar recebendo estas?",
         "mark_all_as_read": "Marcar todas como lidas",
         "mute_channel": "Silenciar canal",
         "muted_2": "silenciado",
         "muted_count": "Silenciado ({{count}})",
-        "no_categories_selected_click_a_category_above": "Nenhuma categoria selecionada — clique em uma categoria acima",
         "no_notifications": "Sem notificações",
         "reject": "Rejeitar",
         "search": "Pesquisar…",
-        "toggle_kind": "Alternar {{kind}}",
-        "select_all_categories": "Selecionar todas as categorias",
         "show_less": "Mostrar menos",
         "try_a_different_search": "Tente outra busca"
       }
