+++ website/src/i18n/locales/ru.json
@@ -4386,17 +4386,13 @@
         "unknown_date": "Дата неизвестна",
         "yesterday": "Вчера",
         "kind_agent": "Агент",
-        "kind_all": "Все",
         "kind_approval": "Подтверждение",
-        "kind_cron": "Cron",
         "kind_cron_job": "Задание cron",
         "kind_heartbeat": "Heartbeat",
-        "kind_hooks": "Хуки",
         "kind_notification": "Уведомление",
         "kind_skills": "Навыки",
         "kind_subagent": "Субагент",
         "kind_task_runner": "Запуск задач",
-        "kind_tasks": "Задачи",
         "kind_webhook": "Вебхук"
       },
       "notificationDetailPanel": {
@@ -4423,24 +4419,18 @@
         "all": "Все",
         "approve": "Подтвердить",
         "clear": "Очистить",
-        "clear_all_filters": "Сбросить все фильтры",
         "clear_all_notifications": "Очистить все уведомления?",
         "dismiss_notification": "Скрыть уведомление",
-        "filter_notifications_by_kind": "Фильтровать уведомления по типу",
         "first_notification_from": "Первое уведомление от",
-        "hide_muted_channel_notifications": "Скрывать уведомления каналов без звука",
         "keep": "Оставить",
         "keep_receiving_these": ". Продолжать получать такие уведомления?",
         "mark_all_as_read": "Отметить все как прочитанные",
         "mute_channel": "Отключить звук канала",
         "muted_2": "без звука",
         "muted_count": "Без звука ({{count}})",
-        "no_categories_selected_click_a_category_above": "Категории не выбраны — нажми на категорию выше",
         "no_notifications": "Нет уведомлений",
         "reject": "Отклонить",
         "search": "Поиск…",
-        "toggle_kind": "Переключить: {{kind}}",
-        "select_all_categories": "Выбрать все категории",
         "show_less": "Свернуть",
         "try_a_different_search": "Попробуй другой запрос"
       }
