+++ website/src/i18n/locales/zh-CN.json
@@ -4221,17 +4221,13 @@
         "unknown_date": "未知日期",
         "yesterday": "昨天",
         "kind_agent": "代理",
-        "kind_all": "全部",
         "kind_approval": "审批",
-        "kind_cron": "Cron",
         "kind_cron_job": "定时任务",
         "kind_heartbeat": "心跳",
-        "kind_hooks": "钩子",
         "kind_notification": "通知",
         "kind_skills": "技能",
         "kind_subagent": "子代理",
         "kind_task_runner": "任务运行器",
-        "kind_tasks": "任务",
         "kind_webhook": "Webhook"
       },
       "notificationDetailPanel": {
@@ -4258,24 +4254,18 @@
         "all": "全部",
         "approve": "批准",
         "clear": "清除",
-        "clear_all_filters": "清除所有筛选条件",
         "clear_all_notifications": "清除所有通知？",
         "dismiss_notification": "忽略通知",
-        "filter_notifications_by_kind": "按类型筛选通知",
         "first_notification_from": "首条通知来自",
-        "hide_muted_channel_notifications": "隐藏已静音频道的通知",
         "keep": "保留",
         "keep_receiving_these": "。要继续接收这类通知吗？",
         "mark_all_as_read": "全部标为已读",
         "mute_channel": "静音频道",
         "muted_2": "已静音",
         "muted_count": "已静音（{{count}}）",
-        "no_categories_selected_click_a_category_above": "未选择任何分类 — 点击上方的分类",
         "no_notifications": "暂无通知",
         "reject": "拒绝",
         "search": "搜索…",
-        "toggle_kind": "切换{{kind}}",
-        "select_all_categories": "全选分类",
         "show_less": "收起",
         "try_a_different_search": "换个搜索词试试"
       }
